+++ sdks/java/io/iceberg/src/main/java/org/apache/beam/sdk/io/iceberg/IcebergTableCreateConfig.java
@@ -19,6 +19,7 @@ package org.apache.beam.sdk.io.iceberg;
 
 import com.google.auto.value.AutoValue;
 import java.util.List;
+import java.util.Map;
 import org.apache.beam.sdk.schemas.Schema;
 import org.apache.iceberg.PartitionSpec;
 import org.checkerframework.checker.nullness.qual.Nullable;
@@ -41,6 +42,9 @@ public abstract class IcebergTableCreateConfig {
   public abstract @Nullable List<String> getPartitionFields();
 
   @Pure
+  public abstract @Nullable Map<String, String> getTableProperties();
+
+  @Pure
   public static Builder builder() {
     return new AutoValue_IcebergTableCreateConfig.Builder();
   }
@@ -50,6 +54,8 @@ public abstract class IcebergTableCreateConfig {
     public abstract Builder setSchema(Schema schema);
 
     public abstract Builder setPartitionFields(@Nullable List<String> partitionFields);
+
+    public abstract Builder setTableProperties(@Nullable Map<String, String> tableProperties);
 
     @Pure
     public abstract IcebergTableCreateConfig build();